[spirv] Provide literal type hint in cast function (#1034)

+++ tools/clang/test/CodeGenSPIRV/ternary-op.cond-op.hlsl
@@ -90,9 +90,7 @@ void main() {
     float a = cond ? 1 : 0;
 
 // CHECK:      [[c_long:%\d+]] = OpSelect %long {{%\d+}} %long_3000000000 %long_4000000000
-// CHECK-NEXT:        {{%\d+}} = OpConvertSToF %float [[c_long]]
-    // TODO: Use OpSConvert to first convert long to int. Then use OpConvertSToF.
-    float c = cond ? 3000000000 : 4000000000;
+    double c = cond ? 3000000000 : 4000000000;
 
 // CHECK:      [[d_int:%\d+]] = OpSelect %uint {{%\d+}} %uint_1 %uint_0
     uint d = cond ? 1 : 0;
